+++ PyricetiFramework/Runtime/Scripts/UI/UIVisibilityStateEnum.cs
@@ -1,0 +1,10 @@
+﻿namespace PyricetiFramework
+{
+  public enum eUIVisibilityState
+  {
+    hidden,
+    visible,
+    fadingIn,
+    fadingOut
+  }
+}